refactor: improve integration_test.go organization by splitting into focused test files
- Split the large integration_test.go file (1011 lines) into multiple focused test files
- Created basic_execution_test.go for core application execution tests
- Created filtering_test.go for file filtering tests
- Created error_handling_test.go for error scenario tests
- Created user_interaction_test.go for user token confirmation tests
- Created audit_log_test.go for audit logging tests
- Modified original file to be a stub that references new test files
- Retained existing xml_integration_test.go, multi_model_test.go, and rate_limit_test.go
- Marked task as complete in TODO.md
